feat: support API Gateway Resource Policy objects
This checks whether the provided `provider.resourcePolicy` is
an object or an array.

When the value is an array (the original behaviour), the array
is used as an array of statements to create a resource policy
document, like this:

```
{
  Version: '2012-10-17',
  Statement: this.serverless.service.provider.resourcePolicy
}
```

When the value is an object, the provided object is used as the
resource policy document.